refactor: simplify PasswordHash() init.
This change streamlines the process for creating a new instance of
`PasswordHash` by making the `hashers` argument optional and setting
the default hasher to `Argon2Hasher` when the user did not otherwise
specify a hasher.

The `PasswordHash.recommended()` class method retains full backwards
compatability with its previous behavior.

The `PasswordHash` object will now raise `HasherNotAvailable` if the
default `Argon2Hasher` is used but the `argon2-cffi` package is not
available.